pyproject.toml
[project]
name = "ai-ocr"
version = "0.2.0"
description = "Searchable PDF generator: Tesseract OCR + Gemini LLM correction + invisible text layer"
requires-python = ">=3.9"
dependencies = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "google-genai>=1.0.0",
    "PyMuPDF>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pyinstaller>=6.0",
]

[project.scripts]
ai-ocr = "ai_ocr.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
